youtube-summary-service/requirements.txt
fastapi
uvicorn
youtube-transcript-api
google-genai
feedparser
python-dotenv